project-livewire/server/requirements.txt
websockets==14.1
google-genai==1.3.0
python-dotenv==1.0.1